requirements.txt
google-generativeai
python-dotenv
requests
pydantic
pytest
